requirements-tracing.txt
# OpenTelemetry dependencies for distributed tracing
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-httpx>=0.42b0
opentelemetry-instrumentation-logging>=0.42b0
opentelemetry-propagator-b3>=1.21.0
